Move preprocessor out of delineation package
#3
I developed a preprocessor to meet the needs of the modified nested set index algorithm, but this class doesn't fit the purpose of this package. Preprocessing is likely a requirement of the library as a whole, and we will like add additional methods (e.g. dropping some fields). The commit moves the preprocessor up to the root of the library.