Fix GeoDataFrame output fails
#3
Anthony caught this error. The previous implementation has a conversion from a dictionary back to a DataFrame. That process was dropping the geometry type information from the DataFrame (i.e. there was a geometry column but it was not the correct type). This meant the MNSI function returned a GeoDataFrame that could not be output due to missing CRS information.